RSE: Set RSE ATU window to device memory
Update the logic in BL1_1 and BL1_2 to initialise the MPU to set the RSE
address for the ATU window to be treated as device memory.

In BL1_1, we only configure the MPU window if required because we are
using a UART from the host system (which requires ATU mapping).
In this case, we configure the whole ATU window to be device memory.
Neither the ATU mapping nor MPU config need to be done again in BL1_2.

In BL1_2, if the configuration has not yet been done in BL1_1, then we
must configure the MPU. We do not, however, need to configure the ATU as
this will have either been done in BL1_1 or not required (either because
we are not using logging or we are not using a UART from the host
system).

Note as part of this change, compile definition for TFM_PXN_ENABLE to
platform_region_defs so that it will get passed to all components and
not just the runtime.

Change-Id: Id02cb4b28113455bfb8c456dc15bfa496d574459
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>