RSE: Update routing tables format
Update the routing tables TGF file format and the routing tables
themselves to have a single routing table for both send and receive.

This patch essentially makes each link bi-directional given that we use
the same link for sending to and receiving from another node in the
system.

The TGF files are updated to follow this new format, whereby we specify
the links using the following formulation:

0 1 Node 0 Link 0 Node 1 Link 1

This specifies that nodes 0 and 1 are directly connected and that to
communicate with node 1, node 0 sends/receives data on link 0. To
communicate with node 0, node 1 sends/receives data on link 1.

Change-Id: Ie2e5509223346c7fd3b556d1459e85c81f416d5a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>