RSE: Introduce RSE_COMMS_NUMBER_NODES
Introduce a new CMake configuration variable which specifies the number
of nodes in the RSE communications system. Initially, this is set to the
same number as the number of RSEs in the system (as it is assumed that
the RSE comms is initially only between RSEs). This variable will be
used to add other software components to the RSE comms (such as the
SCP).

Change-Id: I7d59a1aee9b2954dd3ea4342fdf1f12c2771c737
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>