RSE: Optimise rse_get_routing_tables
Previously, this function copied the routing tables into a buffer
provided by the caller. This is not optimal, particularly in the case
that there are a significant number of nodes in the system.

Instead, given that the routing tables are constant, we can provide a
pointer to them via this function. In the case of the routing tables
being in the OTP, we need to read them into a buffer but there is no
need to do that more than once. We can therefore stash them in a buffer
and return a pointer to that to the caller.

Change-Id: I02e72a17994a3951bbe4cdc46a5a173ab717f458
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>