RSE: Initialise UART ATU when logging disabled
Ensure that BL1 always initialises the ATU region for the UART, even
when logging in BL1 is disabled. This is required as later components
might enabled logging and therefore use the UART via the ATU.

This patch therefore adds the ATU region initialisation in BL1_2 using
the same ifdefs as initialising the ATU window to device memory.

Change-Id: I324d24234866cd080c291a22aaf4309ef71e9f82
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>