RSE: Fix routing tables OTP read alignment
Reads from the OTP must be 4 byte aligned and the buffers we read into a
multiple of 4 bytes. Previously, the RSE routing tables getter functions
allowed the called to pass in a buffer which had the size matching the
number of nodes in the system (which currently equals the number of
RSEs). This is not necessarily a multiple of 4 and therefore the call to
read from the OTP fails. Fix this by rounding the size of the buffer and
setting the error conditions in the getter functions accordingly.

Change-Id: I92f1430959e3faaa6493f4d337bef02edb7462a4
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>