RSE: Add COD structure definition
Add a definition for the structure layout of the Chip Output Data (COD)
and use it in the CM OTP layout structure. This data contains the CMAC
and root attestation public key.

In future these fields will be generated as part of the CM provisioning
but support for that does not currently exist.

Change-Id: I70be9beeb8c899eb0fb18ad37a2d550490cc3c4c
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>