RSE: Add IAK/RAK certificate handling library
Add a new library which contains the definition of the IAK certificate
by compounding C structures. This definition has fixed size based on the
expected format of the IAK certificate, but can be easily modified
should the format change.

The certificate definition (in DER format) is static and therefore
increases the use of the data size, but this allows for easy reproducing
of it, and removes the requirement for additional libraries (which would
probably increase code size more).

The API allows callers to generate the complete certificate by passing
in the missing parameters, and then ask the library to verify the
signature. The fetch function gets a pointer to the certificate and is
intended to eventually be used by the runtime firmware when the
certificate is required.

Change-Id: I3fa764fba1f96f041f63b41739d1577255fe1706
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>