RSE: Add endorsement certificate provisioning blob
Add support for the endorsement provisioning flow in the existing
default_plain_data_handler. This function is already used in the plain
data handler blob and therefore allows for provisioning the endorsement
certificate in the same manner as the non-endorsed provisioning.

Note that there are two possible options for running the plain data blob
which parses the endorsement certificate data - DM LCS in the ROM or SE
LCS in the runtime. This patch only initially adds support for running
the blob in DM in the ROM.

Change-Id: I37d703b155c38ecd3c5aab4ff38662e3d5f54647
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>