RSE: Use ASN1 library for ECDSA key DER generation
Make the IAK certificate generation library generic by renaming the
files and definitions to rse_asn1_. We then add a new function to this
library which generates the ECDSA public key DER. This makes use of the
existing structure definitions and makes the code much simpler by
instead copying an instance of the structure into a given buffer.

We then make use of this new function in the provisioning code, in place
of the manual generation of the key in DER format.

Change-Id: I9e01e4ae8a3b68ad19e11e58781a04eb6c9f8c0a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>