RSE: Add persistent data flags area
Add persistent data flags region to persistent data for storing flags
/bitfields. These are stored in dedicated GPAON registers that persist
over resets if available. If there are insufficient registers available
, then instead the flags are stored in the normal persistent data region
in SRAM.

If the register is available its address should be defined in
RSE_PERSISTENT_DATA_FLAGS_GPAON_BASE and NUM_OF_PERSISTENT_FLAG_REGS
should be set to the number of contiguous available registers to use.

Getter and setter inline functions are defined to access persistent
data flags. These are agnostic to whether or not registers or SRAM is
used for the persistent flag storage.

(Refactored by gautham.ravichandran@arm.com)

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I2b561a6001ea3b2cf88c2e03526f767861094cac