RSE: BL1-1 DCU setup
Add function to setup DCU vector as early as possible based on whether a
certified debug vector is provided. This replaces the previous mechanism
that used shared runtime storage with a system directly using the
persistent data to simplify from the ROM side.

This also checks against the DCU Mask to ensure that an invalid debug
vector is not written to DCU.

When debug is opened all hardware keys are invalidated to ensure they
cannot be used to sign keys when debugging.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I0dfea1083f1aca1ff28125d03f9c4ed7efda7b8b