RSE: Add rse_soc_uid.c to runtime build
This file is required in the case that the DCSU is compiled in (via
RSE_ENABLE_DCSU_PROVISIONING_COMMS) and the SoC UID is enabled in the
OTP (RSE_OTP_HAS_SOC_AREA). Note the we do not make this compile
conditional on the flag. Instead the functions defined in
the file are conditionally defined depending on the value of the flag.

Change-Id: I7462ba47d1dfb3eddbe249e288430ed0750675f3
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>