RSE: Add endorsement cert provisioning in runtime
Add support for the runtime provisioning partition to receive a plain
data blob which will then be used to receive the endorsement certificate
values. This blob is handled in a similar way to the late-DM
provisioning blob in that the runtime provisioning partition HAL resets
the system and allows BL1_1 to process/run the blob. This patch also
adds support for BL1_1 processing the blob.

Note that as part of this change, we increase the runtime provisioning
partition blob buffer. This is required as the plain data blob which is
used to receive the endorsement cert values is much larger than the
existing late-DM provisioning blob as it must contain IAK certificate
generation logic.

Change-Id: I9b1ee8b82c5dbbdc5f27bc73b453ce7fcc9d9729