RSE: Add RSE comms platform definitions
Add RSE comms platform definitions required by the new design. This
includes the TGF file specifying the system layout as well as the
required getter functions which map the node ID to the device structure
pointer.

This patch adds these definitions for each of the RSE platforms that
require them, namely TC, RDV3 and RD1AE.

Change-Id: If66c051280a3c7c7fd1fae2bed0c846296a69134
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>