RSE: Blob provisioning update sp mode handling
Add ability to switch from secure provisioning (SP) enable to SP disable
via reset if blob does not require SP enable but it was already enabled.

Add additional hardening using FIH_COND_CHECK* to sp handling.

Remove disabling of debug when SP is not enabled.

SP enable may be left on after a blob has finished if it does not reset.
If another blob is loaded after this it would have errored if it did not
require SP enable.

Disabling of debug is removed as it could enter a boot loop where debug
is disabled here then re-enabled by startup code after the reset.
Disabling debug at this stage is not very helpful anyway as the function
to disable debug could be skipped by the debugger with no possible
recourse in code.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I6b968fd755eb011bfdfa1fd7a43070a7113c93c6