RSE: Move runtime blob provisioning to BL1_2
In the case of booting in DM LCS or performing endorsement certificate
provisioning, the runtime can stash a blob to be unpacked by the boot
loaders. Move the logic which allows this to happen into BL1_2 (instead
of BL1_1).

This means that the only provisioning which is now performed in BL1_1 is
the CM and (if boot in DM LCS is disabled), the DM.

Change-Id: Ie74afc3b8645709c3682043e769db43e79183e68