RSE: Tests: Only setup master key when required
Setting up the master key (via the rse_setup_master_key function), uses
the KTRL key for derivation. This has a finite number of uses
(controlled with an OTP counter). Given that we continually setup and
tear-down the key in the test suite, we can reach the limit causing the
tests to fail.

Add a static variable to only initialise the key when required. Note
that this variable will not persist on reset, which works well for this
use-case as we need to setup the master key on each reset anyway.

Change-Id: I054ddb43474d0a83b3c336ba01ab4245defd23df
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>