RSE: Add minimal BL1-1 OTP init
In BL1-1 when BL1-2 is going to be loaded, the whole OTP does not need
to be verified in BL1-1, as it will be verified again in BL1-2.

Instead, in order to enter BL1-2 earlier, only BL1-2 OTP region is zero
counted in minimal init as this is required to ensure that BL1-2 has not
been tampered.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I3b72670028f0690b8219dd14739d8dd89ba51a45