RSE: Integrate RSE comms into RSE handshake
Make use of the new RSE comms API in the RSE handshake module. This
patch also adds the RSE comms source files into the BL1_1 build so that
they can be used by the RSE handshake module in BL1_2.

Note that initially, the RSE comms API does not support encryption and
therefore we keep the existing encryption implementation in the RSE
handshake module and only encrypt the handshake message as previously.

Change-Id: I1d6585baca64bccf778c9a6df29474acf061ebb1
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>