tools: switch to deterministic ECDSA signing
non-deterministic ECDSA signing relies on generating a random nonce for
each ECDSA signature. This implies a reliance on the quality of the RNG
source used, and makes the signing process unreproducible.

Instead, switch to deterministic signing, in which the nonce is derived
from the private key as well as the message hash ; as specified in RFC
6979 [0].

[0]: https://datatracker.ietf.org/doc/html/rfc6979.html

Change-Id: Ic66a26b236aaba6c41df639fcdd29b5c45d757d2
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>