RSE: Add MHU device struct version field
Previously, there was nothing stopping us passing a MHUv3 device
structure into the MHUv2 driver wrapper and vice-versa. As the top level
API takes in a void pointer, a pointer to either structure will be
passed in and we will only find out there is an issue when function
returns as error due to the differing register definitions.

Add a new field to the MHUv2 and MHUv3 device structures which is set to
the type of MHU that the structure defines. This is places in the first
field of the structure so that whether the void pointer is cast to the
v2 or v3 structure type, dereferencing it will get the value of the
version field correctly. We then check the value of this field and
return an error (no assert) so that the caller can handle this if
required.

Note that in general, we have a PLAT_MHU_VERSION flag which ensures that
we only compile in structure definitions for a single MHU version. There
are platforms, however, which contain MHUs of different versions and
both of the structure definitions can be compiled in simultaneously.
This check is primary targeted at those platforms, but it acts as a
reasonable sanity check for other platforms.

Change-Id: I1ad08e71d77a0a6e5a9073346536d933dd5c99a2
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>