cc3xx: initialise temp_point during EC point multiplication
When DPA countermeasures are disabled, temp_point MAY be negated at the
end of point multiplication. Since temp_point was never initialized, it
may be mapped to a random PKA buffer which may have a value greater
than the field modulus, in which case, the PKA would assert and fail
during PKA mod_neg operation. To avoid such scenario, initialize temp_point to 0.

Change-Id: I89298470e655f3d63588751e295e1b7e4a79e125
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>