Platform: Fix shared area initialisation
`tlv_tot_len` was incorrectly initialised to `data_size`, which
represents the total size of the shared memory. This caused the
following condition to be evaluated as always `true` and cause
`measured_boot_init` to fail (if CONFIG_TFM_BOOT_STORE_MEASUREMENTS
is ON)
```
else if ((SHARED_DATA_ENTRY_SIZE(size) + boot_data->header.tlv_tot_len) > data_size) {
    return -1;
}
```

Similarly, offset was initialised to the end of the buffer, instead of
being set the end of the added shared buffer.

Change-Id: Iecd88a65e2b1f2073619ba5e959de7ec76015480
Signed-off-by: Mudit Sharma <mudit.sharma@arm.com>