Crypto: fix out-of-bound access in mbedtls PSA code.
In case key_length is equal to block_size, memset would be called with a
destination address outside the boundary of ipad and opad, with a null
size. Since memset implementation is compiler-dependent, it is a good
practice not to assume that it won't dereference the out-of-bound
destination address.

Instead, only call memset if key_length is less than block_size.

Change-Id: Ia36a9816031e02ed1f0cedb03c1f77199a2dde0c
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>