RSE: Specify registers for startup variables
Specifying the specific registers to use for the variables ensures that
they are indeed stored in these registers (the register keyword is just
a suggestion to the compiler). If the compiler cannot place the values
in these registers, then the compilation will fail.

Note that there is nothing special about the chosen registers. We just
use the standard scratch registers.

Change-Id: I5546d4257e393e1fb835d8c838271740c8446fa6
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>