Platform: SAM: Fix SAMICV register writes
Previously, we calculated the zero count of the new register value and
subtracted it from the old register value before adding that to the
value of the SAMICV register and writing that back to the SAMICV
register.

In the case that there are fewer zeros in the new value than the old
value. Then the subtraction would become negative and because these
values are unsigned we would be adding a large positive value to the
SAMICV register. Instead of this, split out the SAMICV writes into their
own function and compare the values of the zero counts before writing.

Change-Id: I134b6f4de8ea70d7e78fb417ac50cd7699f064d0
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>