RSE: Add comms handler support
Add support for RSE comms handler mode. This is the interrupt driven
mode in the RSE comms design. This patch adds the top level handler mode
API which allows applications to register handlers with a specific
application ID. This handler will be called when a message is received
which is associated with this ID. The handler can then queue or process
the message immediately, and will pop the message/reply from the
internal buffer using the appropriate functions.

This patch also adds means for the underlying interrupt handler to
return the allocated buffer to the higher level API as well as pop the
buffer once it has been processed.

Change-Id: Id87388dc92b23b3025046a1c8c118c9968ad3fd3
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>