RSE: Split out comms protocol error
Split out the definition of rse_comms_protocol_error_t into a separate
header file. This allows it to be used by other modules then the top
level RSE comms API. Note that we do not move the definition into the
rse_comms.h header as it does not need to be exposed publicly.

Change-Id: I1bac6b4cb400fe254e7ca7893778b0b53c8e1ba3
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>