FIH: Add support for fih based condition checking
Add macros that can be used in conditional statements to provide FIH
protection to branching conditions.

The protection is in the form of performing the condition check 3 times
with delay between such that if one check is skipped the other checks
will still occur.

FIH_COND_CHECK_SAFE_SKIP does this by using && to re-check only if
the condition is true. This can ensure that a branch is not entered
but will not prevent the branch being wrongly skipped. This can be used
if there is code that should only be run if a condition is true.

FIH_COND_CHECK_SAFE_ENTER is the opposite and will protect against
skipping a branch but not against entering the condition if wrongly
false. This can be used when there is conditional code that is safer to
run than to skip.

FIH_COND_CHECK is safe in both directions but is more complex. Using
 both && and || it will only enter or skip the branch if all three
checks are the same. If there is any inconsistency it will enter a panic
state to prevent either branch being taken. This is for use when taking
the wrong branch would be dangerous in either direction.

If FIH is disabled the condition will pass through unmodified.

If it would be dangerous to call 'a' when not needed but safe to skip
when wanted (e.g. opening debug) then use:
if (FIH_COND_CHECK_SAFE_SKIP(cond)) {
    a();
}

If it would be dangerous to skip calling 'a' when needed but safe to
call it when not needed (e.g. checking for an error) then use:
if (FIH_COND_CHECK_SAFE_ENTER(cond)) {
    a();
}

If there is a branching condition where both paths are dangerous if the
condition is wrong then use:
if (FIH_COND_CHECK(cond)) {
    a();
}
else {
    b();
}

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: I5a29671a839efcb49170079cf88ee98281805bb3