readthedocs: Restore sphinx installation via requirements
With the recent introduction of uv for python package management,
the steps to build the docs were updated as well.
Turns out that some steps are still missing for full functional
environment (error appears to be due to python not being found).

While we find the proper solution, restore the readthedocs sphinx
installation via requirements, as previously.

Fixes: e88101a00 ("Python: Introduce python configuration with pyproject.toml")

Signed-off-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: Ibd6fd94fd66e646d421f9b3a198788ea090269aa