Build: Armclang linker script fix
Fix ITCM and DTCM placement in the Armclang scatter file. The global
location counter ('.') does not exist in Armclang; it is a GNU ld (GCC)
concept.

Therefore, in case of ITCM, the S_RUNTIME_ITCM_START region is placed to
the end of the code part of the Load Region. In the current solution,
ER_SPM_CODE appears after ER_TFM_CODE and ER_RO_DATA. This works,
because the Armclang linker will still pick tfm_spm* for ITCM even if
ER_SPM_CODE is after ER_TFM_CODE. Armclang resolves multiple matches using
specificity, not text order. The *tfm_spm* selector is more specific
than the fallback *(+RO-CODE) and *(+RO-DATA), so tfm_spm symbols are
assigned to ER_SPM_CODE even if that ER appears later in the file.

Accordingly, the DTCM related placements are moved to the very end of
the Load Region.

Signed-off-by: G. Varga, Gabor <Gabor.G.Varga@arm.com>
Signed-off-by: Meggyesi, Janos <janos.meggyesi@arm.com>
Change-Id: I4c5a0f4238cabe33b9c884eac0a5c9c1ad31969c