RSE: Use OTPW or reset_syndrome for init flag
Replace the current usage of the GRETREG for persisting the boot flag
across resets with the RSE persistent data module. This module uses the
existing flags in the OTPW GP_AON registers and also adds functionality
to use the RESET_SYNDROME when the GP_AON registers are not present.

This patch also refactors the inline functions into macros to allow them
to be used safely within the startup_bl1_1.c function which does not use
the stack

Change-Id: Ifa68ca39f3b4daeda7cd5a6fa16892952a30640a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>