RSE: Share the MPU driver from BL1-1 into BL1-2
Previously, the MPU driver was directly compiled in platform_bl1_2
as it was called during disable_rom_execution(), i.e. once the ROM
execution is disabled the driver could not be shared. But as now
we have modified disable_rom_execution() to avoid calling the MPU
driver and directly deal with MPU registers, we can just share the
MPU driver from BL11.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I9160e13d7af3406e145dde80e5ba0504848cee31