RSE: Add comms encryption support
Update the format of the cryptography metadata and use it to encrypt and
decrypt packets. This patch adds a specific BL implementation of the
encryption/decryption as this makes use of the BL1 specific APIs.
Initially, the runtime implementation are just stubs.

We also add support for session key handshakes. This is done
transparently to the caller of the receive/send message functions.

Change-Id: I82a0f0705d141d35e946604e25e5d172b5b462ac
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>