platform: nordic: Update the Nordic USART function
Update the Nordic USART functions to accept other
miscellaneous control commands apart from the
mode control ones.

This makes the USART functions for Nordic platforms
to acept the CONTROL_TX and CONTROL_RX commands as
well.

The sdtio_init function uses the ARM_USART_CONTROL_TX
command which used to fail for Nordic platforms
before but the return code of the control function
was never checked. Recently this changed and it now
checks for the return code which makes the Nordic
platform failing to boot so this change is necessary.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I65b0884338df924dbec3783421779d64ceaaf741