RSE: SFCP: Add support for re-keying
Add support in the RSE comms encryption to re-key the session key in the
case that it reaches a predefined value. This is added to the existing
RSE comms handshake implementation and is designed to work in exactly
the same way.

Initially, both the initial handshake and re-key are hardcoded to be
synchronous on the requester side.

Note that, as part of this commit, we add locking whenever calling any
of the handle functions for the message we are receiving. This prevents
other nodes from communicating with the current processing node which it
is sending a handshake message or updating its internal state.

Change-Id: I343923db99a9cce3e1fd41137b5e1f17fffc6ada
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>