Build: Updated version reading mechanism
When multiple tags are merged into main branch, the current git describe shows the latest. This may not be the correct one, say v2.1.3 versus 2.2.1.
Thus amend the "git describe" command to pick the most recently created tag in history from the branch.

Taggers need to be aware of the order of which tags are created, so that the one created last is the one that will be shown (which is relevant only with merge tags).

Signed-off-by: Anton Komlev <anton.komlev@arm.com>
Co-authored-by: Nicola Mazzucato <nicola.mazzucato@arm.com>
Change-Id: I04b8a4116fab445e3d6819e1700fa82c58b99e93