Build: common: Reuse copy sections for SP stacks
When BL2 loads the secure runtime image, it copies all of the data in
the FLASH memory region to SRAM. This memory region includes elf
sections like .TFM_DATA which is then subsequently copied to a
different memory region of SRAM. When the image is created, copied
sections take up space in the FLASH section which is then unused after
startup. We can minimize the size of the secure runtime code memory
region (which is the FLASH linker script region) by overlaying the
runtime stacks with the copied data.

The change:
- Reorganizes the SP data regions, .TFM_DATA and .SPM_CODE to be
  adjacent and extends the area to to be able to fit all the SP stacks
- Adds all the SP stacks into a separate section and overlaps it to the
  previously created area.
- Also adds an MPU region for this area
- If this feature is enabled SPM clears the stacks before usage.

Change-Id: Iff1120bc2b3bd1960bd5c4bfde59094c182ebbd9
Signed-off-by: Gabor Toth <gabor.toth2@arm.com>