rse: Add script to generate BL2 code encryption keys
This improves using encrypted images as now code
encryption keys are generated as part of the build
and used to encrypt images which are then decrypted
on the RSE using the same key derivation mechanism
so the user doesn't have extract keys from RSE.

Also update default keys in MCUBoot to match what's
in key_enc.c so that encrypted image works in case
of non-RSE platforms that uses keys embedded in BL2
image.

Also remove dependency on BL1_2 in key derivation for
SIC and KCE keys.

Change-Id: Ibd3ad8187440e2157bff82faaccf761ea5e017dd
Signed-off-by: walelm01 <waleed.elmelegy@arm.com>