RSE: Make platform comms size macro generic
Change the macro defined by RSE subplatforms to instead define the total
permitted size of the comms payload instead of the maximum size of the
embedded PSA message. We then use this new macro to determine the
maximum permitted size of the embedded PSA message given size of the
rest of the message metadata.

Change-Id: I9710be02b4883e2bd0021a5e4bbc3f88f9631c86
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>