platform: nordic_nrf: Configure NRF53 RPMSG NS region
The nRF5340 uses the RPMSG region for communication
between the application and the network core over IPC.
The network core is configured as non secure so in order
to allow communication between the two cores this region
needs to be configured as non secure as well.

This adds the configuration of the region if the partition
manager defined a region for it.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I3821967e83dacbd100fecd9f6afa1b3230910c0d