Encrypted image: get rid of some warnings related to configs
This include surrounding functions in thin psa crypto
only used in cipher and key wrap APIs which are only
used if Encrypted image is disabled.

Remove "#cmakedefine MCUBOOT_ENC_IMAGES" from
mcuboot_config.h.in as this was only local to this file
and replaced with
"$<$<BOOL:${MCUBOOT_ENC_IMAGES}>:MCUBOOT_ENC_IMAGES>" added
in CmakeLists file in an earlier patch to avoid multiple
definitions.

Change-Id: Ia1fab39965d04c655594d9ec7292d3edcabbb8d4
Signed-off-by: walelm01 <waleed.elmelegy@arm.com>