RSE: mhu_send_data(): fix MHUv2 note on buffer length
The current note instructs the caller to pad the buffer with 4 bytes when
"size" is divisible by 4 -- the padding size is prescribed with the
expression (4 - (size % 4)). Contrarily, when "size" is a whole multiple
of 4, then 0 bytes of padding are needed; see "mhu_wrapper_v2_x.c":

>     p = (uint32_t *)send_buffer;
>     for (i = 0; i < size; i += 4) {
>         err = mhu_v2_x_channel_send(dev, chan, *p++);

For rounding up the buffer length properly, update the expression to
((size + 3) & ~(size_t)0b11).

Fixes: https://github.com/TrustedFirmware-M/trusted-firmware-m/issues/38
Change-Id: Ie0c7de70d56c80a3169fcd6cadbd7d8a0f1648d4
Signed-off-by: Laszlo Ersek <laszlo.ersek@arm.com>