RSE: SFCP: Add support for mutual authentication
Add support for the mutual authentication handshake in the case that the
trusted subnet has initially untrusted links. This performs most of the
same session key setup as with untrusted links and then sends a mutual
authentication message to each of the client nodes which is then
decrypted with the derived key. The client then sends an empty reply
which the server again decrypts to authenticate the client.

After this process, we can upgrade the link to trusted and no further
encryption is required.

Change-Id: Id798abffef8ef8ee1850719648097e468c8d1142
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>