RSE: Fix send IVs message lookup
Previously, we were comparing the element in the array with the index
corresponding to the current device's node ID. With non-contiguous node
IDs, this would not work correctly. We therefore update the logic to
instead match that used when sending the message (in that the IVs are
sent in the node order given in the trusted subnet configuration
structure.

Change-Id: I7870ac8f804055d0b3bbbe098f892e40c9c2055e
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>