cc3xx: psa: simplify and optimize cc3xx_hash_compute()
The implementation currently relies on using a multi-part operation in
order to compute a hash, which is slow and requires space to save and
restore context between each step of the multi-part operation.

Instead, greatly simplify the process by directly initializing the
cc3xx HASH engine, feeding the input, digest the output and cleanup.

Change-Id: If597016c2e77cc9dc9b3fb028c76bfd2553d09e9
Signed-off-by: Amjad Ouled-Ameur <amjad.ouled-ameur@arm.com>