RSE: Use backing store for SAMICV calculation
Instead of using the existing SAMICV value to calculate the new value,
use a backing store for the registers protected by the SAMICV. This is
updated whenever one of the registers is written and then the zero count
calculated from all of the stored register values.

Note that the existing behaviour is maintained, in that if the event
counter is not present, we do not attempt to write to its shadow
register or include its value in the SAMICV calculation.

Change-Id: I86456d57d77555ce26ce8d4aa49abefe529ae3c5
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>