RSE: SKU: Feature Control: OTP-based DCU override
This change adds feature control to cm provisioning bundle
The OTP-based DCU override field specifies the number of the
FEATURE_CONTROL bits to be copied to the specific DCU line.
The feature is disabled if the field value is 0.
The non-blocking dcsu handler is extended to return the
received command to make post-processing possible.

Change-Id: If07f3745fbe04c38f6c7675752d75f12cf13efc1
Signed-off-by: Gabor Toth <gabor.toth2@arm.com>